Split response_matrix and added config tests
